Update README.md
These scripts are provided as-is. You accept sole responsibility for any and all outcome(s) of their use, of course. I make no claims that some or all of the code is original. Some (il)logic or parts thereof are my work, however.

If something is here, I've used it at one time or another in the past.

They're intended for enhancing some aspect of your

Cheers !